ethportal-api: Refactor portal content keys and values architecture (#806)
* Refactor portal content keys architecture.

* Refactor portal content values architecture.
