+++ ethportal-api/src/beacon.rs
@@ -1,11 +1,11 @@
-use crate::types::content_key::BeaconContentKey;
-use crate::types::content_value::{BeaconContentValue, PossibleBeaconContentValue};
+use crate::types::content_key::beacon::BeaconContentKey;
 use crate::types::enr::Enr;
 use crate::types::portal::FindNodesInfo;
 use crate::types::portal::{
     AcceptInfo, ContentInfo, DataRadius, PaginateLocalContentInfo, PongInfo, TraceContentInfo,
 };
 use crate::RoutingTableInfo;
+use crate::{BeaconContentValue, PossibleBeaconContentValue};
 use discv5::enr::NodeId;
 use jsonrpsee::{core::RpcResult, proc_macros::rpc};
