+++ ethportal-api/src/lib.rs
@@ -19,15 +19,21 @@ pub use beacon::{BeaconNetworkApiClient, BeaconNetworkApiServer};
 pub use history::{HistoryNetworkApiClient, HistoryNetworkApiServer};
 pub use web3::{Web3ApiClient, Web3ApiServer};
 
-// Re-exports trin-types
 pub use types::content_key::{
-    BeaconContentKey, BlockBodyKey, BlockHeaderKey, BlockReceiptsKey, EpochAccumulatorKey,
-    HistoryContentKey, LightClientBootstrapKey, LightClientUpdatesKey, OverlayContentKey,
-    StateContentKey,
+    beacon::{BeaconContentKey, LightClientBootstrapKey, LightClientUpdatesKey},
+    error::ContentKeyError,
+    history::{
+        BlockBodyKey, BlockHeaderKey, BlockReceiptsKey, EpochAccumulatorKey, HistoryContentKey,
+        RawContentKey,
+    },
+    overlay::{IdentityContentKey, OverlayContentKey},
+    state::StateContentKey,
 };
+
 pub use types::content_value::{
-    BeaconContentValue, ContentValue, ContentValueError, HistoryContentValue,
-    PossibleBeaconContentValue, PossibleHistoryContentValue,
+    beacon::{BeaconContentValue, PossibleBeaconContentValue},
+    error::ContentValueError,
+    history::{HistoryContentValue, PossibleHistoryContentValue},
 };
 pub use types::execution::block_body::*;
 pub use types::execution::header::*;
@@ -35,6 +41,7 @@ pub use types::execution::receipts::*;
 
 // Re-exports jsonrpsee crate
 pub use jsonrpsee;
+pub use types::content_value::ContentValue;
 
 pub use types::discv5::*;
 pub use types::enr::*;
