+++ ethportal-api/src/types/content_value/constants.rs
@@ -1,0 +1,3 @@
+/// The length of the Merkle proof for the inclusion of a block header in a particular epoch
+/// accumulator.
+pub const EPOCH_ACC_PROOF_LEN: usize = 15;
