+++ ethportal-api/src/types/portal.rs
@@ -1,6 +1,6 @@
-use crate::types::content_key::HistoryContentKey;
-use crate::types::content_value::{HistoryContentValue, PossibleHistoryContentValue};
 use crate::types::enr::Enr;
+use crate::HistoryContentKey;
+use crate::{HistoryContentValue, PossibleHistoryContentValue};
 use serde::{Deserialize, Serialize};
 use ssz_types::{typenum, BitList};
