+++ newsfragments/806.changed.md
@@ -1,0 +1,1 @@
+ethportal-api: Refactor portal content keys and values architecture
