+++ portalnet/src/find/query_info.rs
@@ -6,8 +6,8 @@ use crate::{
     find::query_pool::TargetKey,
     types::messages::{FindContent, FindNodes, Request},
 };
-use ethportal_api::types::content_key::OverlayContentKey;
 use ethportal_api::types::query_trace::QueryTrace;
+use ethportal_api::OverlayContentKey;
 
 /// Information about a query.
 #[derive(Debug)]
