+++ trin-history/src/jsonrpc.rs
@@ -1,14 +1,16 @@
 use std::sync::Arc;
 
 use discv5::enr::NodeId;
+use ethportal_api::types::content_value::ContentValue;
 use ethportal_api::types::{
-    constants::CONTENT_ABSENT, content_key::RawContentKey, jsonrpc::endpoints::HistoryEndpoint,
+    constants::CONTENT_ABSENT, jsonrpc::endpoints::HistoryEndpoint,
     jsonrpc::request::HistoryJsonRpcRequest, query_trace::QueryTrace,
 };
 use ethportal_api::utils::bytes::hex_encode;
+use ethportal_api::OverlayContentKey;
 use ethportal_api::{
     types::portal::{AcceptInfo, FindNodesInfo, PongInfo, TraceContentInfo},
-    ContentValue, {HistoryContentKey, OverlayContentKey},
+    HistoryContentKey, RawContentKey,
 };
 use portalnet::storage::ContentStore;
 use serde_json::{json, Value};
