+++ trin-validation/src/accumulator.rs
@@ -16,12 +16,12 @@ use crate::{
     constants::{EPOCH_SIZE, MERGE_BLOCK_NUMBER},
     merkle::proof::{verify_merkle_proof, MerkleTree},
 };
-use ethportal_api::types::content_key::{EpochAccumulatorKey, HistoryContentKey};
 use ethportal_api::types::execution::accumulator::EpochAccumulator;
 use ethportal_api::types::execution::header::{BlockHeaderProof, Header, HeaderWithProof};
 use ethportal_api::types::jsonrpc::endpoints::HistoryEndpoint;
 use ethportal_api::types::jsonrpc::request::HistoryJsonRpcRequest;
 use ethportal_api::utils::bytes::hex_decode;
+use ethportal_api::{EpochAccumulatorKey, HistoryContentKey};
 
 /// SSZ List[Hash256, max_length = MAX_HISTORICAL_EPOCHS]
 /// List of historical epoch accumulator merkle roots preceding current epoch.
