+++ trin-validation/src/oracle.rs
@@ -5,11 +5,11 @@ use ssz::Decode;
 use tokio::sync::mpsc;
 
 use crate::accumulator::MasterAccumulator;
-use ethportal_api::types::content_key::{BlockHeaderKey, HistoryContentKey};
 use ethportal_api::types::execution::header::HeaderWithProof;
 use ethportal_api::types::jsonrpc::endpoints::HistoryEndpoint;
 use ethportal_api::types::jsonrpc::request::{BeaconJsonRpcRequest, HistoryJsonRpcRequest};
 use ethportal_api::utils::bytes::hex_decode;
+use ethportal_api::{BlockHeaderKey, HistoryContentKey};
 
 /// Responsible for dispatching cross-overlay-network requests
 /// for data to perform validation.
